refactor: useToolbox change structure
