--- front/src/utils/editorjs/EditorJsContainer/useToolbox/appendItems.js
@@ -1,7 +1,0 @@
-const appendItems = ({ node, items = [] }) => {
-  items.forEach((item) => {
-    node?.appendChild(item);
-  });
-};
-
-export default appendItems;
