--- front/src/utils/editorjs/EditorJsContainer/useToolbox/getBasicAndInteractiveItems.js
@@ -1,11 +1,0 @@
-const getBasicAndInteractiveItems = (items, interactiveItemsName) => {
-  const interactiveItems = items.filter((block) =>
-    interactiveItemsName.includes(block.dataset.tool),
-  );
-  const basicItems = items.filter(
-    (block) => !interactiveItemsName.includes(block.dataset.tool),
-  );
-  return [basicItems, interactiveItems];
-};
-
-export default getBasicAndInteractiveItems;
