--- front/src/utils/editorjs/EditorJsContainer/useToolbox/getToolboxBlocks.js
@@ -1,4 +1,0 @@
-const getToolboxBlocks = (parent) =>
-  parent?.querySelectorAll('.ce-toolbox__button') || [];
-
-export default getToolboxBlocks;
