--- front/src/utils/editorjs/EditorJsContainer/useToolbox/getTranslationKey.js
@@ -1,12 +1,0 @@
-const getTranslationKey = (name) => {
-  switch (name) {
-    case 'fillTheGap':
-      return 'fill_the_gap';
-    case 'closedQuestion':
-      return 'closed_question';
-    default:
-      return name;
-  }
-};
-
-export default getTranslationKey;
