--- front/src/utils/editorjs/EditorJsContainer/useToolbox/updateInnerText.js
@@ -1,8 +1,0 @@
-const updateInnerText = ({ parentNode, text, selector }) => {
-  const element = parentNode.querySelector(selector);
-  if (element) {
-    element.innerText = text;
-  }
-};
-
-export default updateInnerText;
