+++ front/src/utils/editorjs/EditorJsContainer/useToolbox/useToolbox.js
@@ -2,15 +2,19 @@ import { useCallback, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import config from '@sb-ui/utils/api/config';
-import appendItems from '@sb-ui/utils/editorjs/EditorJsContainer/useToolbox/appendItems';
 
-import createDivWithClassName from './createDivWithClassName';
-import getBasicAndInteractiveItems from './getBasicAndInteractiveItems';
-import getTitleKeys from './getTitleKeys';
-import getToolboxBlocks from './getToolboxBlocks';
-import getTranslationKey from './getTranslationKey';
-import selectBlocksDescKeys from './selectBlockDescKeys';
-import updateInnerText from './updateInnerText';
+import {
+  appendItems,
+  createDivWithClassName,
+  getToolboxBlocks,
+  transformDefaultMenuItems,
+  updateInnerText,
+} from './domToolboxHelpers';
+import {
+  getBasicAndInteractiveItems,
+  getTitleKeys,
+  selectBlocksDescKeys,
+} from './toolboxItemsHelpers';
 
 const { interactiveBlocks } = config;
 
@@ -33,50 +37,6 @@ export const useToolbox = () => {
       }))
       .forEach(updateInnerText);
   }, [t]);
-
-  const createItemData = useCallback(
-    (toolName) =>
-      createDivWithClassName({
-        className: 'toolbox-block-data',
-        items: [
-          createDivWithClassName({
-            className: 'toolbox-block-data-name',
-            innerText: t(`tools.${toolName}.title`),
-          }),
-          createDivWithClassName({
-            className: 'toolbox-block-data-description',
-            innerText: t(`tools.${toolName}.description`),
-          }),
-        ],
-      }),
-    [t],
-  );
-
-  const appendItemsToBlock = useCallback(
-    (items, block) => {
-      items.forEach((item) => {
-        block.appendChild(item);
-        const blockItem = items.find(
-          (x) => x.dataset.tool === item.dataset.tool,
-        );
-        const translateToolName = getTranslationKey(item.dataset.tool);
-        appendItems({
-          node: blockItem,
-          items: [
-            createDivWithClassName({
-              className: 'toolbox-svg-wrapper',
-              items: [blockItem.querySelector('svg')],
-            }),
-            createDivWithClassName({
-              className: 'toolbox-block-wrapper',
-              items: [createItemData(translateToolName)],
-            }),
-          ],
-        });
-      });
-    },
-    [createItemData],
-  );
 
   const prepareToolbox = useCallback(() => {
     toolbox.current = document.querySelector('.ce-toolbox');
@@ -115,9 +75,9 @@ export const useToolbox = () => {
       interactiveBlocks,
     );
 
-    appendItemsToBlock(interactiveItems, interactiveBlocksWrapper);
-    appendItemsToBlock(basicItems, basicBlocksWrapper);
-  }, [appendItemsToBlock, t]);
+    transformDefaultMenuItems(interactiveItems, interactiveBlocksWrapper, t);
+    transformDefaultMenuItems(basicItems, basicBlocksWrapper, t);
+  }, [t]);
 
   return {
     prepareToolbox,
